fixed 500 status errors

+++ passenger-service/src/main/java/com/modsen/passengerservice/dto/PassengerResponse.java
@@ -14,5 +14,5 @@ public class PassengerResponse {
     String surname;
     String phone;
     String email;
-    Float rating;
+    Double rating;
 }
